Learning about classes and instances
Java is an object-oriented programming (OOP) language, which means that we can design classes, objects, and methods that can perform certain actions. These behaviors are important in the construction of larger, more robust Java programs.

In this lesson, we will explore some fundamental concepts of object-oriented programming to take advantage of the power of OOP in Java.